Do not aggressively unlock; unlock after use
Here's a scenario where aggressive unlocking may conflict with retries on timeout:

- Tx is sent, times out (and server is not successful at sending the tx out). It will retry 5 seconds after the library raises a timeout exception
- Assuming a few seconds later all the channels have been locked, other jobs will aggressively unlock channels
- When the channel that was used is unlocked, and successfully processes another transaction, when the retry occurs, it will force lock and use the same channel. The sequence number has now been bumped up, and the retry will get a duplicate sequence number error.

- Do not aggressively unlock since we don't know what other channels are locked for timeout reasons. Unlock a channel after use, since we know that we're done with it.
- Have another marker that will remove disallow channels from aggressively being unlocked

Although the latter is more efficient, navigating the locking/unlocking logic will become more difficult. Furthermore, I do not put it beyond myself to have overlooked some corner case. Since creating channels is fairly cheap, we can just increase the lock expiry time and instruct the users of the gem to have more channels. Another concession is to eventually decrease the lock time to a time that is certainly higher than timeouts, but lower than the setting now.

https://github.com/bloom-solutions/stellar_spectrum-ruby/pull/9#issuecomment-446508118

Unlock the channel after use
