+++ lib/stellar_spectrum.rb
@@ -7,7 +7,6 @@ require "redis"
 require "active_support/core_ext/hash/except"
 require "active_support/core_ext/object/blank"
 require 'active_support/core_ext/integer/inflections'
-require "stellar_spectrum/services/get_channel_account_info"
 require "stellar_spectrum/services/get_available_channels"
 require "stellar_spectrum/services/get_channel_accounts"
 require "stellar_spectrum/services/get_key_for_address"
@@ -19,13 +18,9 @@ require "stellar_spectrum/services/init_stellar_client"
 require "stellar_spectrum/services/lock_channel"
 require "stellar_spectrum/services/pick_channel"
 require "stellar_spectrum/services/send_payment"
-require "stellar_spectrum/services/sending_payment/attempt_release_lock"
 require "stellar_spectrum/services/sending_payment/retry"
 require "stellar_spectrum/services/sending_payment/send_asset"
-require "stellar_spectrum/services/unlocking/attempt_release"
-require "stellar_spectrum/services/unlocking/check_sequence_number"
-require "stellar_spectrum/services/unlocking/get_account_to_unlock"
-require "stellar_spectrum/services/unlocking/unlock"
+require "stellar_spectrum/services/unlock"
 
 module StellarSpectrum
